update Controllers, fix update methods.

+++ EshopBackend/src/main/java/EshopProject/EshopBackend/service/ProductServiceImpl.java
@@ -8,6 +8,7 @@ import org.springframework.stereotype.Service;
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Objects;
 import java.util.Optional;
 
 @Service
@@ -44,14 +45,56 @@ public class ProductServiceImpl implements ProductService{
         return productRepository.findById(productId);
     }
 
-//    TBD
     @Override
     public Product updateProduct(Product product, Long productId) {
-        return null;
+
+    Optional<Product> optionalProduct = productRepository.findById(productId);
+
+    if (optionalProduct.isEmpty()) {
+        throw new RuntimeException("Product with ID " + productId + " not found");
     }
+
+    Product currentProduct = optionalProduct.get();
+
+    if (Objects.nonNull(product.getProduct_name()) && !"".equalsIgnoreCase(product.getProduct_name())) {
+        currentProduct.setProduct_name(product.getProduct_name());
+    }
+
+    if (Objects.nonNull(product.getProduct_description()) && !"".equalsIgnoreCase(product.getProduct_description())) {
+        currentProduct.setProduct_description(product.getProduct_description());
+    }
+
+    if (Objects.nonNull(product.getProduct_category()) && !"".equalsIgnoreCase(product.getProduct_category())) {
+        currentProduct.setProduct_category(product.getProduct_category());
+    }
+
+    if (Objects.nonNull(product.getProduct_price())) {
+        currentProduct.setProduct_price(product.getProduct_price());
+    }
+
+    if (Objects.nonNull(product.getProduct_sale_price())) {
+        currentProduct.setProduct_sale_price(product.getProduct_sale_price());
+    }
+
+    if (Objects.nonNull(product.getProduct_stock())) {
+        currentProduct.setProduct_stock(product.getProduct_stock());
+    }
+
+    if (Objects.nonNull(product.getProduct_reviews()) && !"".equalsIgnoreCase(product.getProduct_reviews())) {
+        currentProduct.setProduct_reviews(product.getProduct_reviews());
+    }
+
+    // Save the updated product back to the database
+    return productRepository.save(currentProduct);
+}
+
 
     @Override
     public void deleteProductById(Long productId) {
-        productRepository.deleteById(productId);
+        if (productRepository.existsById(productId)) {
+            productRepository.deleteById(productId);
+        } else {
+            throw new RuntimeException("Product with id " + productId + " not found");
+        }
     }
 }
